project management Jira integration update
