+++ corporate_services/fixtures/workspace.json
@@ -1459,9 +1459,9 @@
     "hidden": 0,
     "icon": null,
     "is_query_report": 0,
-    "label": "Project Progress Report",
+    "label": "Weekly Progress Report",
     "link_count": 0,
-    "link_to": "Project Progress Report",
+    "link_to": "Weekly Progress Report",
     "link_type": "DocType",
     "onboard": 0,
     "only_for": null,
