+++ corporate_services/icl_corporate_services/custom/task.json
@@ -11,6 +11,70 @@
    "collapsible": 0,
    "collapsible_depends_on": null,
    "columns": 0,
+   "creation": "2026-07-23 12:00:00.000000",
+   "default": null,
+   "depends_on": "eval:doc.custom_task_source=='Jira'",
+   "description": "Jira assignee's email, kept here when it doesn't match any Employee's User ID.",
+   "docstatus": 0,
+   "dt": "Task",
+   "fetch_from": null,
+   "fetch_if_empty": 0,
+   "fieldname": "custom_jira_assignee_email",
+   "fieldtype": "Data",
+   "hidden": 0,
+   "hide_border": 0,
+   "hide_days": 0,
+   "hide_seconds": 0,
+   "idx": 19,
+   "ignore_user_permissions": 0,
+   "ignore_xss_filter": 0,
+   "in_global_search": 0,
+   "in_list_view": 0,
+   "in_preview": 0,
+   "in_standard_filter": 0,
+   "insert_after": "custom_jira_issue_url",
+   "is_system_generated": 0,
+   "is_virtual": 0,
+   "label": "Jira Assignee Email",
+   "length": 0,
+   "link_filters": null,
+   "mandatory_depends_on": null,
+   "modified": "2026-07-23 12:00:00.000000",
+   "modified_by": "Administrator",
+   "module": null,
+   "name": "Task-custom_jira_assignee_email",
+   "no_copy": 0,
+   "non_negative": 0,
+   "options": "Email",
+   "owner": "Administrator",
+   "permlevel": 0,
+   "placeholder": null,
+   "precision": "",
+   "print_hide": 0,
+   "print_hide_if_no_value": 0,
+   "print_width": null,
+   "read_only": 1,
+   "read_only_depends_on": null,
+   "report_hide": 0,
+   "reqd": 0,
+   "search_index": 0,
+   "show_dashboard": 0,
+   "sort_options": 0,
+   "translatable": 0,
+   "unique": 0,
+   "width": null
+  },
+  {
+   "_assign": null,
+   "_comments": null,
+   "_liked_by": null,
+   "_user_tags": null,
+   "allow_in_quick_entry": 0,
+   "allow_on_submit": 0,
+   "bold": 0,
+   "collapsible": 0,
+   "collapsible_depends_on": null,
+   "columns": 0,
    "creation": "2025-06-17 10:37:39.128286",
    "default": null,
    "depends_on": null,
@@ -32,7 +96,7 @@
    "in_list_view": 0,
    "in_preview": 0,
    "in_standard_filter": 0,
-   "insert_after": "custom_jira_issue_url",
+   "insert_after": "custom_jira_assignee_email",
    "is_system_generated": 0,
    "is_virtual": 0,
    "label": "Allocate To",
@@ -75,6 +139,70 @@
    "collapsible": 0,
    "collapsible_depends_on": null,
    "columns": 0,
+   "creation": "2026-07-23 12:30:00.000000",
+   "default": null,
+   "depends_on": "eval:doc.custom_task_source=='Jira'",
+   "description": "The Jira sprint this issue was last part of, if any.",
+   "docstatus": 0,
+   "dt": "Task",
+   "fetch_from": null,
+   "fetch_if_empty": 0,
+   "fieldname": "custom_jira_sprint",
+   "fieldtype": "Link",
+   "hidden": 0,
+   "hide_border": 0,
+   "hide_days": 0,
+   "hide_seconds": 0,
+   "idx": 19,
+   "ignore_user_permissions": 0,
+   "ignore_xss_filter": 0,
+   "in_global_search": 0,
+   "in_list_view": 0,
+   "in_preview": 0,
+   "in_standard_filter": 0,
+   "insert_after": "custom_allocate_to",
+   "is_system_generated": 0,
+   "is_virtual": 0,
+   "label": "Jira Sprint",
+   "length": 0,
+   "link_filters": null,
+   "mandatory_depends_on": null,
+   "modified": "2026-07-23 12:30:00.000000",
+   "modified_by": "Administrator",
+   "module": null,
+   "name": "Task-custom_jira_sprint",
+   "no_copy": 0,
+   "non_negative": 0,
+   "options": "Jira Sprint",
+   "owner": "Administrator",
+   "permlevel": 0,
+   "placeholder": null,
+   "precision": "",
+   "print_hide": 0,
+   "print_hide_if_no_value": 0,
+   "print_width": null,
+   "read_only": 1,
+   "read_only_depends_on": null,
+   "report_hide": 0,
+   "reqd": 0,
+   "search_index": 0,
+   "show_dashboard": 0,
+   "sort_options": 0,
+   "translatable": 0,
+   "unique": 0,
+   "width": null
+  },
+  {
+   "_assign": null,
+   "_comments": null,
+   "_liked_by": null,
+   "_user_tags": null,
+   "allow_in_quick_entry": 0,
+   "allow_on_submit": 0,
+   "bold": 0,
+   "collapsible": 0,
+   "collapsible_depends_on": null,
+   "columns": 0,
    "creation": "2026-06-12 08:49:36.353942",
    "default": null,
    "depends_on": null,
@@ -96,7 +224,7 @@
    "in_list_view": 0,
    "in_preview": 0,
    "in_standard_filter": 0,
-   "insert_after": "custom_allocate_to",
+   "insert_after": "custom_jira_sprint",
    "is_system_generated": 0,
    "is_virtual": 0,
    "label": "",
@@ -730,7 +858,7 @@
    "property": "field_order",
    "property_type": "Data",
    "row_name": null,
-   "value": "[\"subject\", \"project\", \"issue\", \"type\", \"color\", \"is_group\", \"is_template\", \"column_break0\", \"status\", \"priority\", \"task_weight\", \"parent_task\", \"completed_by\", \"completed_on\", \"custom_task_origin\", \"custom_task_source\", \"custom_jira_issue_key\", \"custom_jira_issue_url\", \"custom_allocate_to\", \"custom_column_break_wqrov\", \"custom_email_sent_date\", \"custom_send_email\", \"custom_email_sent\", \"custom_last_milestone_alert_date\", \"sb_timeline\", \"exp_start_date\", \"expected_time\", \"start\", \"column_break_11\", \"exp_end_date\", \"progress\", \"duration\", \"is_milestone\", \"sb_details\", \"description\", \"sb_depends_on\", \"depends_on\", \"depends_on_tasks\", \"sb_actual\", \"act_start_date\", \"actual_time\", \"column_break_15\", \"act_end_date\", \"sb_costing\", \"total_costing_amount\", \"total_expense_claim\", \"column_break_20\", \"total_billing_amount\", \"sb_more_info\", \"review_date\", \"closing_date\", \"column_break_22\", \"department\", \"company\", \"lft\", \"rgt\", \"old_parent\", \"template_task\"]"
+   "value": "[\"subject\", \"project\", \"issue\", \"type\", \"color\", \"is_group\", \"is_template\", \"column_break0\", \"status\", \"priority\", \"task_weight\", \"parent_task\", \"completed_by\", \"completed_on\", \"custom_task_origin\", \"custom_task_source\", \"custom_jira_issue_key\", \"custom_jira_issue_url\", \"custom_jira_assignee_email\", \"custom_allocate_to\", \"custom_jira_sprint\", \"custom_column_break_wqrov\", \"custom_email_sent_date\", \"custom_send_email\", \"custom_email_sent\", \"custom_last_milestone_alert_date\", \"sb_timeline\", \"exp_start_date\", \"expected_time\", \"start\", \"column_break_11\", \"exp_end_date\", \"progress\", \"duration\", \"is_milestone\", \"sb_details\", \"description\", \"sb_depends_on\", \"depends_on\", \"depends_on_tasks\", \"sb_actual\", \"act_start_date\", \"actual_time\", \"column_break_15\", \"act_end_date\", \"sb_costing\", \"total_costing_amount\", \"total_expense_claim\", \"column_break_20\", \"total_billing_amount\", \"sb_more_info\", \"review_date\", \"closing_date\", \"column_break_22\", \"department\", \"company\", \"lft\", \"rgt\", \"old_parent\", \"template_task\"]"
   }
  ],
  "sync_on_migrate": 1
