+++ corporate_services/icl_corporate_services/doctype/communication_method/communication_method.js
@@ -1,0 +1,8 @@
+// Copyright (c) 2026, IntelliSOFT Consulting and contributors
+// For license information, please see license.txt
+
+// frappe.ui.form.on("Communication Method", {
+// 	refresh(frm) {
+
+// 	},
+// });
