+++ corporate_services/icl_corporate_services/doctype/communication_method/communication_method.py
@@ -1,0 +1,9 @@
+# Copyright (c) 2026, IntelliSOFT Consulting and contributors
+# For license information, please see license.txt
+
+# import frappe
+from frappe.model.document import Document
+
+
+class CommunicationMethod(Document):
+	pass
