+++ corporate_services/icl_corporate_services/doctype/communication_method/test_communication_method.py
@@ -1,0 +1,9 @@
+# Copyright (c) 2026, IntelliSOFT Consulting and Contributors
+# See license.txt
+
+# import frappe
+from frappe.tests.utils import FrappeTestCase
+
+
+class TestCommunicationMethod(FrappeTestCase):
+	pass
