+++ corporate_services/icl_corporate_services/doctype/jira_sprint/jira_sprint.json
@@ -1,0 +1,113 @@
+{
+ "actions": [],
+ "allow_rename": 1,
+ "autoname": "field:sprint_id",
+ "creation": "2026-07-23 19:30:00.000000",
+ "doctype": "DocType",
+ "engine": "InnoDB",
+ "field_order": [
+  "sprint_id",
+  "sprint_name",
+  "jira_project",
+  "column_break_main",
+  "state",
+  "board_id",
+  "dates_section",
+  "start_date",
+  "end_date",
+  "column_break_dates",
+  "goal"
+ ],
+ "fields": [
+  {
+   "fieldname": "sprint_id",
+   "fieldtype": "Data",
+   "label": "Sprint ID",
+   "reqd": 1,
+   "unique": 1
+  },
+  {
+   "fieldname": "sprint_name",
+   "fieldtype": "Data",
+   "in_list_view": 1,
+   "label": "Sprint Name"
+  },
+  {
+   "fieldname": "jira_project",
+   "fieldtype": "Link",
+   "in_list_view": 1,
+   "in_standard_filter": 1,
+   "label": "Jira Project",
+   "options": "Jira Project"
+  },
+  {
+   "fieldname": "column_break_main",
+   "fieldtype": "Column Break"
+  },
+  {
+   "fieldname": "state",
+   "fieldtype": "Select",
+   "in_list_view": 1,
+   "in_standard_filter": 1,
+   "label": "State",
+   "options": "\nFuture\nActive\nClosed"
+  },
+  {
+   "fieldname": "board_id",
+   "fieldtype": "Data",
+   "label": "Board ID"
+  },
+  {
+   "fieldname": "dates_section",
+   "fieldtype": "Section Break",
+   "label": "Dates"
+  },
+  {
+   "fieldname": "start_date",
+   "fieldtype": "Date",
+   "in_list_view": 1,
+   "label": "Start Date"
+  },
+  {
+   "fieldname": "end_date",
+   "fieldtype": "Date",
+   "in_list_view": 1,
+   "label": "End Date"
+  },
+  {
+   "fieldname": "column_break_dates",
+   "fieldtype": "Column Break"
+  },
+  {
+   "fieldname": "goal",
+   "fieldtype": "Small Text",
+   "label": "Goal"
+  }
+ ],
+ "index_web_pages_for_search": 1,
+ "links": [],
+ "modified": "2026-07-23 19:30:00.000000",
+ "modified_by": "Administrator",
+ "module": "ICL Corporate Services",
+ "name": "Jira Sprint",
+ "owner": "Administrator",
+ "permissions": [
+  {
+   "create": 1,
+   "delete": 1,
+   "email": 1,
+   "export": 1,
+   "print": 1,
+   "read": 1,
+   "report": 1,
+   "role": "System Manager",
+   "share": 1,
+   "write": 1
+  }
+ ],
+ "row_format": "Dynamic",
+ "sort_field": "modified",
+ "sort_order": "DESC",
+ "states": [],
+ "track_changes": 1
+}
