+++ corporate_services/icl_corporate_services/doctype/jira_sprint/jira_sprint.py
@@ -1,0 +1,8 @@
+# Copyright (c) 2026, ICL and contributors
+# For license information, please see license.txt
+
+from frappe.model.document import Document
+
+
+class JiraSprint(Document):
+	pass
