+++ corporate_services/icl_corporate_services/print_format/exit_interview/exit_interview.json
@@ -1,7 +1,7 @@
 {
  "absolute_value": 0,
  "align_labels_right": 0,
- "creation": "2026-07-23 11:38:41.724396",
+ "creation": "2026-07-23 19:56:38.140813",
  "css": "  body {\n    font-family: 'EB Garamond', serif;\n    font-size: 11pt;\n    color: #1a1a1a;\n    background: #fff;\n    line-height: 1.6;\n  }\n \n  .page {\n    width: 100%;\n    box-sizing: border-box;\n  }\n \n \n  /* \u2500\u2500 COMPANY LOGO \u2500\u2500 */\n  .company-logo {\n    height: 52px;\n    width: auto;\n    object-fit: contain;\n    display: block;\n    margin-bottom: 6px;\n  }\n  .logo-placeholder {\n    font-family: 'Montserrat', sans-serif;\n    font-size: 9pt;\n    font-weight: 700;\n    letter-spacing: 0.22em;\n    text-transform: uppercase;\n    color: #aaa;\n    border: 1.5px dashed #ccc;\n    padding: 6px 14px;\n    display: inline-block;\n    margin-bottom: 6px;\n  }\n \n  /* \u2500\u2500 HEADER \u2500\u2500 */\n  .doc-header {\n    border-bottom: 2.5px solid #1a1a1a;\n    padding-bottom: 10px;\n    margin-bottom: 20px;\n  }\n  .company-name {\n    font-family: 'Montserrat', sans-serif;\n    font-size: 8pt;\n    font-weight: 700;\n    letter-spacing: 0.2em;\n    text-transform: uppercase;\n    color: #555;\n  }\n  .doc-title {\n    font-family: 'EB Garamond', serif;\n    font-size: 26pt;\n    font-weight: 600;\n    line-height: 1.1;\n    color: #1a1a1a;\n  }\n  .doc-meta {\n    text-align: right;\n    font-family: 'Montserrat', sans-serif;\n    font-size: 8pt;\n    color: #555;\n    line-height: 2;\n    max-width: 45%;\n    word-break: break-word;\n  }\n  .doc-meta .doc-id {\n    font-weight: 700;\n    font-size: 9.5pt;\n    color: #1a1a1a;\n    letter-spacing: 0.04em;\n    display: block;\n  }\n  .status-badge {\n    font-family: 'Montserrat', sans-serif;\n    font-size: 7pt;\n    font-weight: 700;\n    letter-spacing: 0.14em;\n    text-transform: uppercase;\n    border: 1.5px solid #1a1a1a;\n    padding: 2px 9px;\n    color: #1a1a1a;\n    display: inline-block;\n    margin-top: 4px;\n  }\n \n  /* \u2500\u2500 SECTION LABEL \u2500\u2500 */\n  .section-label {\n    font-family: 'Montserrat', sans-serif;\n    font-size: 7.5pt;\n    font-weight: 700;\n    letter-spacing: 0.2em;\n    text-transform: uppercase;\n    color: #555;\n    margin: 20px 0 8px 0;\n    padding-bottom: 5px;\n    border-bottom: 1px solid #ccc;\n  }\n \n  /* \u2500\u2500 EMPLOYEE DETAILS TABLE \u2500\u2500 */\n  .emp-table {\n    width: 100%;\n    border-collapse: collapse;\n    margin-bottom: 4px;\n  }\n  .emp-table th {\n    font-family: 'Montserrat', sans-serif;\n    font-size: 7.5pt;\n    font-weight: 700;\n    text-transform: uppercase;\n    letter-spacing: 0.08em;\n    color: #fff;\n    background: #1a1a1a;\n    padding: 6px 10px;\n    border: 1px solid #1a1a1a;\n    white-space: nowrap;\n    vertical-align: middle;\n  }\n  .emp-table td {\n    font-size: 10.5pt;\n    color: #1a1a1a;\n    padding: 7px 10px;\n    border: 1px solid #ddd;\n    vertical-align: middle;\n  }\n  .emp-table tr:nth-child(even) td { background: #fafafa; }\n \n  /* \u2500\u2500 RATING TABLE \u2500\u2500 */\n  .rating-table {\n    width: 100%;\n    border-collapse: collapse;\n    margin-bottom: 4px;\n  }\n  .rating-table thead th {\n    font-family: 'Montserrat', sans-serif;\n    font-size: 7.5pt;\n    font-weight: 700;\n    text-transform: uppercase;\n    letter-spacing: 0.1em;\n    color: #fff;\n    background: #1a1a1a;\n    padding: 7px 10px;\n    border: 1px solid #1a1a1a;\n  }\n  .rating-table tbody td {\n    padding: 7px 10px;\n    border: 1px solid #ddd;\n    font-size: 10.5pt;\n    vertical-align: middle;\n  }\n  .rating-table tbody tr:nth-child(even) td { background: #fafafa; }\n  .rating-table .rating-val {\n    text-align: center;\n    font-family: 'Montserrat', sans-serif;\n    font-weight: 700;\n    font-size: 11pt;\n  }\n  .rating-table .comment-val {\n    font-style: italic;\n    color: #444;\n    font-size: 10pt;\n  }\n \n  /* \u2500\u2500 Q&A TABLE \u2500\u2500 */\n  .qa-table {\n    width: 100%;\n    border-collapse: collapse;\n    margin-bottom: 4px;\n  }\n  .qa-table td {\n    padding: 7px 10px;\n    border: 1px solid #ddd;\n    vertical-align: top;\n    font-size: 10.5pt;\n  }\n  .qa-table .q-cell {\n    font-family: 'Montserrat', sans-serif;\n    font-size: 8pt;\n    font-weight: 600;\n    color: #333;\n    background: #f5f5f5;\n    width: 36%;\n    line-height: 1.5;\n  }\n  .qa-table .a-cell { color: #1a1a1a; }\n  .qa-table .a-cell p, .qa-table .a-cell span { font-size: 10.5pt; }\n \n  /* \u2500\u2500 SIGNATURE \u2500\u2500 */\n  .sig-section {\n    margin-top: 28px;\n    border-top: 1px solid #ccc;\n    padding-top: 16px;\n  }\n  .sig-line {\n    border-bottom: 1px solid #1a1a1a;\n    height: 30px;\n    margin-bottom: 5px;\n  }\n  .sig-meta {\n    font-family: 'Montserrat', sans-serif;\n    font-size: 7.5pt;\n    color: #666;\n    letter-spacing: 0.05em;\n    text-transform: uppercase;\n  }\n \n  /* \u2500\u2500 FOOTER \u2500\u2500 */\n  .doc-footer {\n    margin-top: 22px;\n    border-top: 1px solid #e0e0e0;\n    padding-top: 8px;\n  }\n  .footer-text {\n    font-family: 'Montserrat', sans-serif;\n    font-size: 7pt;\n    color: #aaa;\n    letter-spacing: 0.08em;\n    text-transform: uppercase;\n  }\n \n  @media print {\n    body { -webkit-print-color-adjust: exact; print-color-adjust: exact; }\n  }",
  "custom_format": 1,
  "default_print_language": "en",
@@ -17,7 +17,7 @@
  "margin_left": 15.0,
  "margin_right": 15.0,
  "margin_top": 15.0,
- "modified": "2026-07-23 11:39:12.073728",
+ "modified": "2026-07-23 19:57:11.215838",
  "modified_by": "Administrator",
  "module": "ICL Corporate Services",
  "name": "Exit Interview",
