+++ corporate_services/icl_corporate_services/print_format/staff_requisition/staff_requisition.json
@@ -1,7 +1,7 @@
 {
  "absolute_value": 0,
  "align_labels_right": 0,
- "creation": "2026-07-23 11:38:40.344278",
+ "creation": "2026-07-23 19:56:36.577141",
  "css": "\n    * {\n        margin: 0;\n        padding: 0;\n        box-sizing: border-box;\n    }\n\n    body {\n        font-family: Arial, sans-serif;\n        font-size: 11pt;\n        color: #333;\n        line-height: 1.6;\n        padding: 20px;\n    }\n\n    /* \u2500\u2500 Header \u2500\u2500 */\n    .header-table {\n        width: 100%;\n        border-collapse: collapse;\n        margin-bottom: 10px;\n    }\n    .header-logo-cell {\n        width: 130px;\n        vertical-align: middle;\n        text-align: left;\n    }\n    .header-title-cell {\n        vertical-align: middle;\n        text-align: center;\n    }\n    .header-spacer-cell {\n        width: 130px;\n    }\n    .logo {\n        max-width: 120px;\n        max-height: 60px;\n    }\n    .doc-title {\n        color: #2c3e50;\n        font-size: 22pt;\n        font-weight: bold;\n        margin-bottom: 4px;\n    }\n    .doc-id {\n        font-size: 12pt;\n        color: #555;\n    }\n\n    .separator {\n        border: none;\n        border-top: 3px solid #2c3e50;\n        margin: 16px 0 20px 0;\n    }\n\n    /* \u2500\u2500 Section Header \u2500\u2500 */\n    .section-header {\n        color: #2c3e50;\n        font-size: 13pt;\n        font-weight: bold;\n        padding-bottom: 6px;\n        margin-top: 24px;\n        margin-bottom: 14px;\n        border-bottom: 2px solid #2c3e50;\n    }\n\n    /* \u2500\u2500 Info Grid (2-col via table) \u2500\u2500 */\n    .grid-table {\n        width: 100%;\n        border-collapse: separate;\n        border-spacing: 8px;\n        margin-bottom: 10px;\n    }\n    .info-box {\n        border: 1px solid #ddd;\n        padding: 10px 12px;\n        border-radius: 4px;\n        background-color: #f9f9f9;\n        vertical-align: top;\n        width: 50%;\n    }\n    .info-box-full {\n        border: 1px solid #ddd;\n        padding: 10px 12px;\n        border-radius: 4px;\n        background-color: #f9f9f9;\n        vertical-align: top;\n    }\n    .info-label {\n        font-weight: bold;\n        color: #2c3e50;\n        font-size: 10pt;\n        display: block;\n        margin-bottom: 4px;\n    }\n    .info-value {\n        color: #444;\n        font-size: 11pt;\n    }\n\n    /* \u2500\u2500 Text Content Block \u2500\u2500 */\n    .text-block {\n        border: 1px solid #ddd;\n        padding: 14px;\n        background-color: #f9f9f9;\n        border-radius: 4px;\n        margin-bottom: 16px;\n        min-height: 60px;\n        font-size: 11pt;\n        color: #444;\n        white-space: pre-wrap;\n    }\n\n    /* \u2500\u2500 Approval Table \u2500\u2500 */\n    .approval-table {\n        width: 100%;\n        border-collapse: collapse;\n        margin-top: 12px;\n        margin-bottom: 30px;\n        font-size: 10pt;\n    }\n    .approval-table th {\n        background-color: #34495e;\n        color: #fff;\n        padding: 9px 10px;\n        text-align: left;\n        border: 1px solid #2c3e50;\n        font-size: 10pt;\n    }\n    .approval-table td {\n        padding: 9px 10px;\n        border: 1px solid #ddd;\n    }\n    .approval-table tr.even td {\n        background-color: #f4f4f4;\n    }\n\n    /* \u2500\u2500 Status Badge \u2500\u2500 */\n    .status-badge {\n        display: inline-block;\n        padding: 3px 12px;\n        border-radius: 12px;\n        font-size: 10pt;\n        font-weight: bold;\n        background-color: #e8f4fd;\n        color: #1a6fa0;\n        border: 1px solid #aad4f0;\n    }\n\n    /* \u2500\u2500 Footer \u2500\u2500 */\n    .footer {\n        margin-top: 50px;\n        padding-top: 14px;\n        border-top: 1px solid #ddd;\n        text-align: left;\n    }\n    .footer-logo {\n        max-width: 110px;\n        max-height: 55px;\n    }",
  "custom_format": 1,
  "default_print_language": "en",
@@ -17,7 +17,7 @@
  "margin_left": 15.0,
  "margin_right": 15.0,
  "margin_top": 15.0,
- "modified": "2026-07-23 11:39:10.934318",
+ "modified": "2026-07-23 19:57:10.022399",
  "modified_by": "Administrator",
  "module": "ICL Corporate Services",
  "name": "Staff Requisition",
