+++ corporate_services/icl_corporate_services/print_format/weekly_progress_report/weekly_progress_report.json
@@ -1,7 +1,7 @@
 {
  "absolute_value": 0,
  "align_labels_right": 0,
- "creation": "2026-07-23 11:38:40.277445",
+ "creation": "2026-07-23 19:56:36.497524",
  "css": "* { box-sizing: border-box; margin: 0; padding: 0; }\nbody, .wpr-wrapper { font-family: 'Inter', 'Segoe UI', Arial, sans-serif; font-size: 12px; color: #1a1a2e; background: #ffffff; line-height: 1.6; }\n.wpr-wrapper { max-width: 900px; margin: 0 auto; padding: 28px 32px; }\n.header-table { width: 100%; border-collapse: collapse; margin-bottom: 0; }\n.header-table td { vertical-align: middle; padding: 0; }\n.logo-cell { width: 120px; }\n.company-logo { max-width: 110px; max-height: 60px; object-fit: contain; }\n.logo-placeholder { font-size: 18px; font-weight: 700; color: #1a1a2e; letter-spacing: 2px; }\n.title-cell { padding-left: 20px; }\n.doc-title { font-size: 20px; font-weight: 700; color: #1a1a2e; letter-spacing: 0.3px; }\n.doc-subtitle { font-size: 11px; color: #6b7280; margin-top: 3px; font-weight: 400; text-transform: uppercase; letter-spacing: 0.8px; }\n.ref-cell { width: 230px; vertical-align: top; }\n.ref-table { width: 100%; border-collapse: collapse; border: 1px solid #e5e7eb; border-radius: 6px; overflow: hidden; }\n.ref-table tr:not(:last-child) td { border-bottom: 1px solid #e5e7eb; }\n.ref-label { background: #f9fafb; padding: 5px 10px; font-weight: 600; color: #374151; font-size: 10.5px; white-space: nowrap; width: 45%; border-right: 1px solid #e5e7eb; }\n.ref-value { padding: 5px 10px; color: #1a1a2e; font-size: 10.5px; }\n.badge-status { display: inline-block; padding: 2px 8px; border-radius: 20px; font-size: 10px; font-weight: 600; text-transform: uppercase; letter-spacing: 0.5px; background: #fef9c3; color: #854d0e; border: 1px solid #fde68a; }\n.badge-submitted, .badge-approved { background: #dcfce7; color: #166534; border-color: #bbf7d0; }\n.badge-rejected { background: #fee2e2; color: #991b1b; border-color: #fecaca; }\n.badge-cancelled { background: #f3f4f6; color: #6b7280; border-color: #e5e7eb; }\n.section-divider { height: 2px; background: linear-gradient(to right, #1a1a2e, #4b5563, #e5e7eb); margin: 18px 0 16px 0; border-radius: 2px; }\n.section-label { font-size: 10.5px; font-weight: 700; text-transform: uppercase; letter-spacing: 1.2px; color: #ffffff; background: #1a1a2e; padding: 6px 14px; margin-top: 14px; margin-bottom: 0; border-radius: 4px 4px 0 0; }\n.info-table, .answers-table { width: 100%; border-collapse: collapse; border: 1px solid #e5e7eb; border-top: none; margin-bottom: 0; }\n.info-table tr:not(:last-child) td, .answers-table tr:not(:last-child) td { border-bottom: 1px solid #e5e7eb; }\n.info-key { background: #f9fafb; padding: 8px 14px; font-weight: 600; color: #374151; font-size: 11px; width: 18%; border-right: 1px solid #e5e7eb; white-space: nowrap; vertical-align: top; }\n.info-val { padding: 8px 14px; color: #1a1a2e; font-size: 11px; width: 32%; border-right: 1px solid #e5e7eb; vertical-align: top; }\n.info-val:last-child { border-right: none; }\n.question-cell { background: #f9fafb; padding: 10px 14px; width: 34%; vertical-align: top; border-right: 1px solid #e5e7eb; }\n.question-title { font-size: 11px; font-weight: 700; color: #1a1a2e; }\n.question-help { font-size: 10px; color: #6b7280; margin-top: 4px; }\n.required-mark { color: #b91c1c; font-weight: 700; }\n.response-cell { padding: 10px 14px; color: #1a1a2e; font-size: 11.5px; line-height: 1.7; vertical-align: top; }\n.response-cell p { margin-bottom: 6px; }\n.response-cell ol, .response-cell ul { padding-left: 20px; margin: 4px 0; }\n.response-cell li { margin-bottom: 4px; }\n.empty-response { color: #9ca3af; font-style: italic; }\n.signature-table { width: 100%; border-collapse: collapse; margin-top: 12px; }\n.sig-cell { vertical-align: top; text-align: center; width: 30%; }\n.sig-spacer { width: 5%; }\n.sig-line { height: 1px; background: #1a1a2e; margin-bottom: 8px; }\n.sig-name { font-size: 11px; font-weight: 700; color: #1a1a2e; }\n.sig-role { font-size: 10px; color: #6b7280; margin-top: 2px; }\n.sig-date { font-size: 10px; color: #9ca3af; margin-top: 6px; }\n.footer-table { width: 100%; border-collapse: collapse; border-top: 1px solid #e5e7eb; margin-top: 20px; padding-top: 8px; }\n.footer-table td { padding: 8px 0; font-size: 9.5px; color: #9ca3af; }\n.footer-left { text-align: left; }\n.footer-right { text-align: right; }\n@media print { .wpr-wrapper { padding: 16px 20px; } .section-label { -webkit-print-color-adjust: exact; print-color-adjust: exact; } .header-table, .info-table, .answers-table, .signature-table, .footer-table { page-break-inside: avoid; } }",
  "custom_format": 1,
  "default_print_language": "en",
@@ -17,7 +17,7 @@
  "margin_left": 15.0,
  "margin_right": 15.0,
  "margin_top": 15.0,
- "modified": "2026-07-23 11:39:10.846903",
+ "modified": "2026-07-23 19:57:09.951151",
  "modified_by": "Administrator",
  "module": "ICL Corporate Services",
  "name": "Weekly Progress Report",
