+++ corporate_services/icl_corporate_services/workspace/icl_procurement_&_assets/icl_procurement_&_assets.json
@@ -5,8 +5,8 @@
    "label": "Purchase Order Status"
   }
  ],
- "content": "[{\"id\":\"u8jzPde0Ig\",\"type\":\"header\",\"data\":{\"text\":\"<span class=\\\"h4\\\">ICL Procurement &amp; Assets</span>\",\"col\":12}},{\"id\":\"xLd6GncfBA\",\"type\":\"shortcut\",\"data\":{\"shortcut_name\":\"Supplier\",\"col\":3}},{\"id\":\"epfJBd0Kh8\",\"type\":\"shortcut\",\"data\":{\"shortcut_name\":\"Request for Quotation\",\"col\":3}},{\"id\":\"oOOL8dKLzd\",\"type\":\"shortcut\",\"data\":{\"shortcut_name\":\"Purchase Order\",\"col\":3}},{\"id\":\"ocJ2isAjIh\",\"type\":\"shortcut\",\"data\":{\"shortcut_name\":\"Asset Requisition\",\"col\":3}},{\"id\":\"KtJ0RlgLKO\",\"type\":\"shortcut\",\"data\":{\"shortcut_name\":\"Purchase Order Analysis\",\"col\":3}},{\"id\":\"mxgJTeKdNn\",\"type\":\"shortcut\",\"data\":{\"shortcut_name\":\"Purchase Register\",\"col\":3}},{\"id\":\"FRIBXuDL7D\",\"type\":\"spacer\",\"data\":{\"col\":12}},{\"id\":\"xtpYlSXpfK\",\"type\":\"header\",\"data\":{\"text\":\"<span class=\\\"h4\\\"><b>Key Metrics</b></span>\",\"col\":12}},{\"id\":\"tHF4vUCsMe\",\"type\":\"number_card\",\"data\":{\"number_card_name\":\"Active Suppliers\",\"col\":3}},{\"id\":\"hGAkWvj7FA\",\"type\":\"number_card\",\"data\":{\"number_card_name\":\"Pending RFQs\",\"col\":3}},{\"id\":\"c9QeWJKY40\",\"type\":\"number_card\",\"data\":{\"number_card_name\":\"Purchase Orders to Receive\",\"col\":3}},{\"id\":\"uvSwMFLZDe\",\"type\":\"number_card\",\"data\":{\"number_card_name\":\"Purchase Orders to Bill\",\"col\":3}},{\"id\":\"1f8rESQedU\",\"type\":\"number_card\",\"data\":{\"number_card_name\":\"Total Assets\",\"col\":3}},{\"id\":\"StPKR0CsTy\",\"type\":\"spacer\",\"data\":{\"col\":12}},{\"id\":\"4Qwb8DwkNh\",\"type\":\"header\",\"data\":{\"text\":\"<span class=\\\"h4\\\"><b>Purchase Analytics</b></span>\",\"col\":12}},{\"id\":\"FdnXsiVpzz\",\"type\":\"chart\",\"data\":{\"chart_name\":\"Purchase Order Status\",\"col\":12}},{\"id\":\"63FfkCzJr4\",\"type\":\"spacer\",\"data\":{\"col\":12}},{\"id\":\"i0B3JrTAwR\",\"type\":\"header\",\"data\":{\"text\":\"<span class=\\\"h4\\\"><b>Suppliers &amp; Purchasing</b></span>\",\"col\":12}},{\"id\":\"4y9ojfljoQ\",\"type\":\"card\",\"data\":{\"card_name\":\"Suppliers\",\"col\":3}},{\"id\":\"oaF1Llqsaj\",\"type\":\"quick_list\",\"data\":{\"quick_list_name\":\"Request for Quotation\",\"col\":4}},{\"id\":\"AIxNKu8iS2\",\"type\":\"quick_list\",\"data\":{\"quick_list_name\":\"Supplier Quote Submission\",\"col\":4}},{\"id\":\"G8NPRVdD53\",\"type\":\"spacer\",\"data\":{\"col\":12}},{\"id\":\"X83RZJzzzz\",\"type\":\"card\",\"data\":{\"card_name\":\"Purchase Orders\",\"col\":3}},{\"id\":\"gEOzdmenCk\",\"type\":\"quick_list\",\"data\":{\"quick_list_name\":\"Purchase Order\",\"col\":4}},{\"id\":\"hvMdgaKjIg\",\"type\":\"spacer\",\"data\":{\"col\":12}},{\"id\":\"8xNbe3nNyj\",\"type\":\"card\",\"data\":{\"card_name\":\"Materials & Items\",\"col\":3}},{\"id\":\"Oq9wMxEhh2\",\"type\":\"quick_list\",\"data\":{\"quick_list_name\":\"Material Request\",\"col\":4}},{\"id\":\"FDEEtfjgVv\",\"type\":\"spacer\",\"data\":{\"col\":12}},{\"id\":\"VqE1SkHbn8\",\"type\":\"header\",\"data\":{\"text\":\"<span class=\\\"h4\\\"><b>Asset Management</b></span>\",\"col\":12}},{\"id\":\"8HxjSI6bWH\",\"type\":\"card\",\"data\":{\"card_name\":\"Assets\",\"col\":3}},{\"id\":\"tP3fS2qHx6\",\"type\":\"quick_list\",\"data\":{\"quick_list_name\":\"Asset Requisition\",\"col\":4}},{\"id\":\"kwXoIIXGvO\",\"type\":\"quick_list\",\"data\":{\"quick_list_name\":\"Asset Movement\",\"col\":4}}]",
- "creation": "2026-07-16 13:19:21.873833",
+ "content": "[{\"id\":\"u8jzPde0Ig\",\"type\":\"header\",\"data\":{\"text\":\"<span class=\\\"h4\\\">ICL Procurement &amp; Assets</span>\",\"col\":12}},{\"id\":\"xLd6GncfBA\",\"type\":\"shortcut\",\"data\":{\"shortcut_name\":\"Supplier\",\"col\":3}},{\"id\":\"epfJBd0Kh8\",\"type\":\"shortcut\",\"data\":{\"shortcut_name\":\"Request for Quotation\",\"col\":3}},{\"id\":\"oOOL8dKLzd\",\"type\":\"shortcut\",\"data\":{\"shortcut_name\":\"Purchase Order\",\"col\":3}},{\"id\":\"ocJ2isAjIh\",\"type\":\"shortcut\",\"data\":{\"shortcut_name\":\"Asset Requisition\",\"col\":3}},{\"id\":\"KtJ0RlgLKO\",\"type\":\"shortcut\",\"data\":{\"shortcut_name\":\"Purchase Order Analysis\",\"col\":3}},{\"id\":\"mxgJTeKdNn\",\"type\":\"shortcut\",\"data\":{\"shortcut_name\":\"Purchase Register\",\"col\":3}},{\"id\":\"FRIBXuDL7D\",\"type\":\"spacer\",\"data\":{\"col\":12}},{\"id\":\"xtpYlSXpfK\",\"type\":\"header\",\"data\":{\"text\":\"<span class=\\\"h4\\\"><b>Key Metrics</b></span>\",\"col\":12}},{\"id\":\"tHF4vUCsMe\",\"type\":\"number_card\",\"data\":{\"number_card_name\":\"Active Suppliers\",\"col\":3}},{\"id\":\"hGAkWvj7FA\",\"type\":\"number_card\",\"data\":{\"number_card_name\":\"Pending RFQs\",\"col\":3}},{\"id\":\"c9QeWJKY40\",\"type\":\"number_card\",\"data\":{\"number_card_name\":\"Purchase Orders to Receive\",\"col\":3}},{\"id\":\"uvSwMFLZDe\",\"type\":\"number_card\",\"data\":{\"number_card_name\":\"Purchase Orders to Bill\",\"col\":3}},{\"id\":\"1f8rESQedU\",\"type\":\"number_card\",\"data\":{\"number_card_name\":\"Total Assets\",\"col\":3}},{\"id\":\"StPKR0CsTy\",\"type\":\"spacer\",\"data\":{\"col\":12}},{\"id\":\"4Qwb8DwkNh\",\"type\":\"header\",\"data\":{\"text\":\"<span class=\\\"h4\\\"><b>Purchase Analytics</b></span>\",\"col\":12}},{\"id\":\"FdnXsiVpzz\",\"type\":\"chart\",\"data\":{\"chart_name\":\"Purchase Order Status\",\"col\":6}},{\"id\":\"63FfkCzJr4\",\"type\":\"spacer\",\"data\":{\"col\":12}},{\"id\":\"i0B3JrTAwR\",\"type\":\"header\",\"data\":{\"text\":\"<span class=\\\"h4\\\"><b>Suppliers &amp; Purchasing</b></span>\",\"col\":12}},{\"id\":\"4y9ojfljoQ\",\"type\":\"card\",\"data\":{\"card_name\":\"Suppliers\",\"col\":3}},{\"id\":\"oaF1Llqsaj\",\"type\":\"quick_list\",\"data\":{\"quick_list_name\":\"Request for Quotation\",\"col\":4}},{\"id\":\"AIxNKu8iS2\",\"type\":\"quick_list\",\"data\":{\"quick_list_name\":\"Supplier Quote Submission\",\"col\":4}},{\"id\":\"G8NPRVdD53\",\"type\":\"spacer\",\"data\":{\"col\":12}},{\"id\":\"X83RZJzzzz\",\"type\":\"card\",\"data\":{\"card_name\":\"Purchase Orders\",\"col\":3}},{\"id\":\"gEOzdmenCk\",\"type\":\"quick_list\",\"data\":{\"quick_list_name\":\"Purchase Order\",\"col\":4}},{\"id\":\"hvMdgaKjIg\",\"type\":\"spacer\",\"data\":{\"col\":12}},{\"id\":\"8xNbe3nNyj\",\"type\":\"card\",\"data\":{\"card_name\":\"Materials & Items\",\"col\":3}},{\"id\":\"Oq9wMxEhh2\",\"type\":\"quick_list\",\"data\":{\"quick_list_name\":\"Material Request\",\"col\":4}},{\"id\":\"FDEEtfjgVv\",\"type\":\"spacer\",\"data\":{\"col\":12}},{\"id\":\"VqE1SkHbn8\",\"type\":\"header\",\"data\":{\"text\":\"<span class=\\\"h4\\\"><b>Asset Management</b></span>\",\"col\":12}},{\"id\":\"8HxjSI6bWH\",\"type\":\"card\",\"data\":{\"card_name\":\"Assets\",\"col\":3}},{\"id\":\"tP3fS2qHx6\",\"type\":\"quick_list\",\"data\":{\"quick_list_name\":\"Asset Requisition\",\"col\":4}},{\"id\":\"kwXoIIXGvO\",\"type\":\"quick_list\",\"data\":{\"quick_list_name\":\"Asset Movement\",\"col\":4}}]",
+ "creation": "2026-07-23 11:38:44.915996",
  "custom_blocks": [],
  "docstatus": 0,
  "doctype": "Workspace",
@@ -225,7 +225,7 @@
    "type": "Link"
   }
  ],
- "modified": "2026-07-16 13:19:21.873833",
+ "modified": "2026-07-23 11:48:51.169213",
  "modified_by": "Administrator",
  "module": "ICL Corporate Services",
  "name": "ICL Procurement & Assets",
