+++ corporate_services/icl_corporate_services/workspace/icl_project_manager/icl_project_manager.json
@@ -155,9 +155,9 @@
   {
    "hidden": 0,
    "is_query_report": 0,
-   "label": "Project Progress Report",
+   "label": "Weekly Progress Report",
    "link_count": 0,
-   "link_to": "Project Progress Report",
+   "link_to": "Weekly Progress Report",
    "link_type": "DocType",
    "onboard": 0,
    "type": "Link"
